+++ src/tests.rs
@@ -21,4 +21,23 @@ quickcheck! {
             && sv.is_compact()
             && (0..n_before_compact).all(|i| sv.get(i).is_some())
     }
+
+    fn from_and_extend_and_from_iter(items: Vec<u8>) -> bool {
+        use std::iter::FromIterator;
+
+        let iter_a = items.iter().cloned();
+        let iter_b = items.iter().cloned();
+
+        let sv_a = StableVec::from_iter(iter_a);
+        let sv_b = {
+            let mut sv = StableVec::new();
+            sv.extend(iter_b);
+            sv
+        };
+        let sv_c = StableVec::from(&items);
+
+        sv_a.num_elements() == items.len()
+            && sv_a == sv_b
+            && sv_a == sv_c
+    }
 }