+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `StableVec::into_vec()`
 - `impl<T> Default for StableVec<T>`
 - Added `FromIterator` impl for `StableVec`
+- Added `Extend` impl for `StableVec`
 - Added `Debug` impl for `StableVec` with a fitting example